requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==1.0.0
nltk==3.8.1
scikit-learn==1.3.2
pdfplumber==0.10.3
textstat==0.7.3
